Rename 'codegen' to 'engine', move Clauses
