+++ core/src/main/scala/io/github/scalaquest/core/parsing/engine/Clauses.scala
@@ -1,4 +1,4 @@
-package io.github.scalaquest.core.parsing.codegen
+package io.github.scalaquest.core.parsing.engine
 
 sealed trait Clause {
   def generate: String
